REFACTOR: router 연결된 컴포넌트 page 폴더로 분리
