+++ 00_signup/src/Router.tsx
@@ -1,7 +1,7 @@
 
 import { Routes, Route } from 'react-router-dom'
-import Welcome from './component/welcome'
-import SignUp from './component/signUp'
+import Welcome from './page/welcome'
+import SignUp from './page/signUp'
 
 const Router = () => {
   return (
